Remove ./ngsc catchup command (#44)
The catchup command was designed to fix issues with people pulling changes from master, and then their app not working because of changes made to the project dependencies without them installing the new versions.

However, no one used this command, including myself, and I realized it's abstracting too much. We want people to have to learn how git works, and this command is bad for hiding git from people.