No longer push to GitHub with `./ngsc deploy` (#97)
Now that the master branch is protected, it is impossible to push directly to master. Instead, new code changes may only be introduced via GitHub pull requests that pass CI and are intentionally merged into master.

So, the deploy script's scope changes to simply be pushing whatever is on GitHub's master to also be on Heroku.